Use new ID format
This approach includes the three key components in the ID, making it
easy to reference any object in the system by it's unique components.
This really matters on the front-end where I struggled to piece together
all the different series and site information across the different
models that were coming back.